move tests closer to components under test
